requirements.txt
# Core LLM & NLP
transformers>=4.36.2

accelerate>=0.29.0
torch>=2.2.0
safetensors

# Optional quantization & performance (only if using quantized models)
bitsandbytes
xformers

# Vector DB
faiss-cpu   # For FAISS vector database (CPU version)
sentence-transformers>=2.2.2

# Streamlit frontend
streamlit>=1.33.0

# Utilities
requests
tqdm
numpy==1.24.4
scikit-learn

# Data Parsing
nltk
spacy

# Ontology formats (optional)
rdflib

# Environment & File Handling
python-dotenv
